pyproject.toml
[tool.poetry]
name = "drf-7-cw"
version = "0.1.0"
description = ""
authors = ["Stereo2025 <teneloks@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
celery = "^5.4.0"
redis = "^5.0.4"
stripe = "^9.6.0"
django-celery-beat = "^2.6.0"
pytelegrambotapi = "^4.18.0"
coverage = "^7.5.1"
pytest = "^8.2.0"
pytest-django = "^4.8.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
